Moving composer components to styled components - CloseButton
