+++ packages/composer/composer/components/LinkAttachment.jsx
@@ -4,12 +4,11 @@
 
 import React from 'react';
 import PropTypes from 'prop-types';
-import AppActionCreators from '../action-creators/AppActionCreators';
 import ComposerActionCreators from '../action-creators/ComposerActionCreators';
 import LinkAttachmentTextEditor from '../components/LinkAttachmentTextEditor';
 import LinkAttachmentThumbnailEditor from '../components/LinkAttachmentThumbnailEditor';
 import LinkAttachmentThumbnail from '../components/LinkAttachmentThumbnail';
-import CloseButton from '../components/CloseButton';
+import CloseButton from '../components/styled/CloseButton';
 import A from '../components/styled/A';
 import { AttachmentTypes, LinkAttachmentTextFieldTypes } from '../AppConstants';
 import styles from './css/LinkAttachment.css';
