+++ packages/composer/composer/components/MediaAttachmentThumbnail.jsx
@@ -6,8 +6,7 @@ import { borderRadius } from '@bufferapp/ui/style/borders';
 import { Text } from '@bufferapp/ui';
 import styled from 'styled-components';
 import ComposerActionCreators from '../action-creators/ComposerActionCreators';
-import AppActionCreators from '../action-creators/AppActionCreators';
-import CloseButton from './CloseButton';
+import CloseButton from './styled/CloseButton';
 import Button from './styled/Button';
 import { escapeParens } from '../utils/StringUtils';
 import styles from './css/MediaAttachmentThumbnail.css';
