+++ packages/composer/composer/components/Notification.jsx
@@ -4,7 +4,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { Button } from '@bufferapp/ui';
 import { NotificationTypes } from '../AppConstants';
-import CloseButton from '../components/CloseButton';
+import CloseButton from './styled/CloseButton';
 import styles from './css/Notification.css';
 import NotificationActionCreators from '../action-creators/NotificationActionCreators';
