+++ packages/composer/composer/components/RetweetAttachment.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import CloseButton from '../components/CloseButton';
+import CloseButton from './styled/CloseButton';
 import AppActionCreators from '../action-creators/AppActionCreators';
 import ComposerActionCreators from '../action-creators/ComposerActionCreators';
 import { AttachmentTypes } from '../AppConstants';
