+++ packages/composer/composer/components/styled/CloseButton.jsx
